pyproject.toml
[project]
name = "twinsync"
version = "0.1.0"
description = ""
authors = [
    {name = "Yash Chauhan",email = "yashc9411407582@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "ruff (>=0.12.9,<0.13.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "google-auth (>=2.40.3,<3.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "drf-spectacular[sidecar] (>=0.28.0,<0.29.0)",
    "requests (>=2.32.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
